card and robot page changes (#13)
* Card Adjustment

Changed the description of tom tom and akhil's cards. I also changed the order of the cards to be from oldest to newest members. The ones that joined the same year are in order of age decending.

* krebby boi

changed van kreb's front photo

* krebby update

added the maid dress for HVK on the back of his card

* powerplay description

made the description of the 2022-2023 powerplay robot which I'm proud of. (kingston)

* Update Krebby v2

* Update ourRobot.html

---------

Co-authored-by: Kingsbytheton <157898635+Kingsbytheton@users.noreply.github.com>
Co-authored-by: RobotNatchos <47341210+RobotNatchos@users.noreply.github.com>